requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
httplib2
idna==2.8
kiwisolver==1.1.0
lxml==4.4.1
matplotlib==3.1.1
numpy==1.17.2
oauthlib
pandas==0.25.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pyparsing==2.4.2
python-dateutil==2.8.0
python-docx
pytz==2019.3
PyYAML
requests
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.6
